Fix tests failures for MacOS
This change depends on the fix on slang-gfx side, so we
have to use the latest release which includes the fix.

However, in order to use the latest release of slang, there
will be some change on slang shader because of the initializer
list change on slang.

Also, metal doesn't allow "main" as entry point function name,
so change those to "computeMain", also add logger to print message
to console for MacOs and Linux.